+++ Modules/Auth/database/migrations/2025_11_02_102315_create_jwt_refresh_tokens_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('jwt_refresh_tokens', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->string('token', 255)->unique();
-            $table->string('ip', 45)->nullable();
-            $table->string('user_agent', 255)->nullable();
-            $table->timestamp('revoked_at')->nullable();
-            $table->timestamp('expires_at')->nullable();
-            $table->timestamps();
-        
-            $table->index(['user_id', 'revoked_at']);
-        });        
+        if (! Schema::hasTable('jwt_refresh_tokens')) {
+            Schema::create('jwt_refresh_tokens', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->string('token', 255)->unique();
+                $table->string('ip', 45)->nullable();
+                $table->string('user_agent', 255)->nullable();
+                $table->timestamp('revoked_at')->nullable();
+                $table->timestamp('expires_at')->nullable();
+                $table->timestamps();
+
+                $table->index(['user_id', 'revoked_at']);
+            });
+        }
     }
 
     /**
