+++ Modules/Auth/database/migrations/2025_11_02_102331_create_login_activities_table.php
@@ -11,7 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('login_activities', function (Blueprint $table) {
+        if (!Schema::hasTable('login_activities')) {
+            Schema::create('login_activities', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->string('ip', 45)->nullable();
@@ -20,7 +21,8 @@ return new class extends Migration
             $table->timestamp('logged_in_at')->nullable();
             $table->timestamp('logged_out_at')->nullable();
             $table->timestamps();
-        });        
+            });
+        }
     }
 
     /**
